realocados fluxos para arquivo exclusivo renomeado models_abstract -> behaviors
