+++ sme_terceirizadas/escola/models.py
@@ -7,7 +7,7 @@ from ..cardapio.models import (
     AlteracaoCardapio, GrupoSuspensaoAlimentacao, InversaoCardapio
 )
 from ..dados_comuns.constants import DAQUI_A_30_DIAS, DAQUI_A_7_DIAS
-from ..dados_comuns.models_abstract import (
+from ..dados_comuns.behaviors import (
     Ativavel, Iniciais, Nomeavel, TemChaveExterna,
     TemCodigoEOL)
 from ..inclusao_alimentacao.models import (
